+++ framework/CHANGELOG.md
@@ -4,6 +4,7 @@ Yii Framework 2 Change Log
 2.0.4 under development
 -----------------------
 
+- Enh #3506: Added `\yii\validators\IpAddressValidator` to perform validation of IP addresses and subnets (SilverFire)
 - Bug #7529: Fixed `yii\web\Response::sendContentAsFile()` that was broken in 2.0.3 (samdark)
 - Bug #7603: Fixed escape characters in `FormatConverter` to work with unicode characters (maddoger, cebe)
 - Enh #6895: Added `ignoreCategories` config option for message command to ignore categories specified (samdark)